# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  '/api/users/{user}/avatar':
    post:
      summary: 'Store a new user avatar.'
      operationId: storeANewUserAvatar
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: user
        description: 'The user.'
        example: 1
        required: true
        schema:
          type: integer
...
  '/api/event/delete/{event_id}':
    post:
      summary: ''
      operationId: postApiEventDeleteEvent_id
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: event_id
        description: 'The ID of the event.'
        example: 1
        required: true
        schema:
          type: integer
...
  '/api/event/report/{event_id}':
    post:
      summary: ''
      operationId: postApiEventReportEvent_id
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                participants_count:
                  type: number
                  description: 'Must be at least 1.'
                  example: 4
                average_participant_age:
                  type: number
                  description: 'Must be at least 1.'
                  example: 90
                percentage_of_females:
                  type: number
                  description: 'Must be between 0 and 100.'
                  example: 1
                name_for_certificate:
                  type: string
                  description: 'Must match the regex /^[^ə]*$/u. Must not be greater than 40 characters.'
                  example: sghmayealoekuq
              required:
                - participants_count
                - average_participant_age
                - percentage_of_females
                - name_for_certificate
      security: []
    parameters:
      -
        in: path
        name: event_id
        description: 'The ID of the event.'
        example: 1
        required: true
        schema:
          type: integer
...